Fix format bug in anomaly detector.
The Anomaly detector had a format discrepancy when sending data to the
Flask server which handles the ML models. I altered the form of the data
in the server to reflect that.

I also addressed the null data issue in the preprocessing stage done by the agent. This was due to an issue calculating normalized values and noting that this needed to be
different in training and preprocessing, but also using the same mean and std dev from training. Hence the mean and std deviation used in training (loaded from serialization).

Having fixed how the models used data for their task, I changed the
anomalydashboard to reflect these changes since it is the Facade that
starts the pipeline.

Other minor changes were made like removing redundant code in
networkflowreader and adjusting parameters (batch size) in autoencoder. The latter
was not a functional change but a matter of speeding up training and
trying to improve autoencoder performance.